LayoutTests/platform/ios/imported/w3c/web-platform-tests/dom/events/click-on-absolute-pseudo-expected.txt


FAIL Click on pseudo-element that disappears on mousedown assert_equals: Click event should fire even if pseudo-element disappears on mousedown expected 1 but got 0
